requirements.txt
Flask==2.0.2
Flask-APScheduler==1.12.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
mashumaro==3.0
psycopg2==2.9.3
PyYAML==6.0
snowflake-sqlalchemy==1.3.3
snowplow-tracker==0.10.0
sqlalchemy-utils==0.38.2
